# Conflicts:
#	Concussion Ball/Assets/Scripts/match/NetworkPlayer.cs
#	Concussion Ball/Concussion Ball.thomas
#	thomas/ThomasCore/src/thomas/object/component/physics/MeshCollider.cpp
#	thomas/ThomasCore/src/thomas/resource/Resource.cpp
#	thomas/ThomasCore/src/thomas/resource/Resource.h
#	thomas/ThomasEngine/src/object/component/RenderComponent.cpp
#	thomas/ThomasEngine/src/object/component/physics/Collider.cpp
#	thomas/ThomasEngine/src/resource/Resources.cpp

+++ thomas/ThomasEngine/src/object/component/RenderComponent.cpp
@@ -114,7 +114,7 @@ namespace ThomasEngine
 	}
 	Material^ RenderComponent::GetMaterial(int index)
 	{
-		if ((uint32_t)index < 0 || render->numMeshes() <= (uint32_t)index) return nullptr;
+		if (index < 0 || render->numMeshes() <= index) return nullptr;
 		thomas::resource::Material* nptr = render->GetMaterial(index);
 		Resource^ mat = ThomasEngine::Resources::FindResourceFromNativePtr(nptr);
 		if (mat != nullptr)
@@ -125,7 +125,7 @@ namespace ThomasEngine
 	{
 		if(mat == nullptr)
 			mat = Material::StandardMaterial;
-		if (index < 0 || render->numMeshes() <= (uint32_t)index)
+		if (index < 0 || render->numMeshes() <= index)
 		{
 			Debug::LogWarning("Failed to set material " + mat->Name + " at index: " + index + ", in RenderComponent: " + Name);
 			return;
